requirements.txt
gspread
oauth2client
beautifulsoup4
aiohttp
python-dotenv
